// MiLicenciaCentro/src/app/enums/TipoDocumento.ts
export enum TipoDocumento {
    CedulaCiudadania = 1,
    CedulaExtranjeria = 2,
    TarjetaIdentidad = 3,
    NIT = 4,
    Pasaporte = 5,
    ContrasenaCedulaCiudadania = 10,
    ContrasenaCedulaExtranjeria = 11,
    TarjetaIdentidadRosada = 12,
    PermisoProteccionTemporal=13,
    RegistroCivil = 14,
    PermisoEspecialPermanencia = 15,
    NUIP = 16
}